Changed array to long long int

+++ project_one/main.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int* bubblesort(int arr[], int num_of_elements) {
+long long int* bubblesort(long long int arr[], int num_of_elements) {
     for(int i=1; i<=num_of_elements-1; i++) {
         for(int j=num_of_elements-1; j>=i; j--) {
             if(arr[j] < arr[j-1]) {
@@ -15,7 +15,7 @@ int* bubblesort(int arr[], int num_of_elements) {
     return arr;
 }
 
-void print_array(int arr[], int num_of_elements) {
+void print_array(long long int arr[], int num_of_elements) {
     for (int i=0; i<num_of_elements; i++) {
         cout << arr[i] << " ";
     }
@@ -27,12 +27,12 @@ int main() {
     int num_of_elements;
     cin >> num_of_elements;
 
-    int array[num_of_elements];
+    long long int array[num_of_elements];
     for (int index=0; index<num_of_elements; index++) {
         cin >> array[index];
     }
 
-    int* sorted_array = bubblesort(array, num_of_elements);
+    long long int* sorted_array = bubblesort(array, num_of_elements);
     print_array(sorted_array, num_of_elements);
 
     return 0;
